requirements.txt
asgiref==3.4.0
backports.entry-points-selectable==1.1.0
blib==0.0.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.1
colorama==0.4.4
cryptography==37.0.4
cycler==0.10.0
Deprecated==1.2.13
distlib==0.3.3
dj-database-url==1.0.0
Django==3.2.4
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jo==0.0.2
joblib==1.0.1
jwcrypto==1.3.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mysql-connector-python==8.0.25
numpy==1.21.2
oauth2client==4.1.3
openpyxl==3.0.8
pandas==1.3.2
phonenumbers==8.12.30
Pillow==8.3.2
platformdirs==2.3.0
protobuf==4.21.5
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pynput==1.7.3
pyparsing==2.4.7
Pyrebase4==4.5.0
PySimpleGUI==4.55.1
python-dateutil==2.8.2
python-jwt==3.3.3
pytz==2021.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
urllib3==1.26.11
uvicorn==0.18.2
virtualenv==20.8.0
virtualenvwrapper-win==1.2.7
Werkzeug==2.0.1
whitenoise==6.2.0
wrapt==1.14.1
